+++ components/Code.js
@@ -5,16 +5,22 @@ import javascript from 'react-syntax-highlighter/dist/languages/prism/javascript
 import style from 'react-syntax-highlighter/dist/styles/prism/atom-dark';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faThumbsDown, faThumbsUp } from '@fortawesome/free-solid-svg-icons';
+import RunKit from 'react-runkit';
 
 SyntaxHighlighter.registerLanguage('bash', bash);
 SyntaxHighlighter.registerLanguage('jsx', jsx);
 SyntaxHighlighter.registerLanguage('javascript', javascript);
 
-export default ({ caption, code, language = 'shell', rating = '' }) => {
+export default ({ caption, code, language = 'shell', rating = '', runkit = false }) => {
 
     return (
-        <div className="code">
-            <SyntaxHighlighter language={language} style={style}>{code.trim()}</SyntaxHighlighter>
+        <div className={`code ${runkit ? 'with-runkit' : ''}`}>
+            { runkit &&
+                <RunKit source={code.trim()} />
+            }
+            { runkit ||
+                <SyntaxHighlighter language={language} style={style}>{code.trim()}</SyntaxHighlighter>
+            }
             { caption &&
                 <p>{caption}</p>
             }